Create utils directory.
