+++ src/module/helpers/_module.mjs
@@ -2,6 +2,6 @@ export * from "./dsn.mjs";
 export * from "./hotReload.mjs";
 export * from "./handlebars.mjs";
 export * as macros from "./macros.mjs";
-export * as utils from "./utils.mjs";
+export * as localization from "./localization.mjs";
 export { default as DrawSteelSettingsHandler } from "./settings.mjs";
 export { default as DrawSteelSocketHandler } from "./sockets.mjs";
